Noticket: fix PMD warnings
git-svn-id: https://svn.cesecore.eu/svn/ejbca-restricted/trunk/ejbca@16865 37aeb1de-b14b-0410-9a0e-82b28170d228

+++ src/java/org/cesecore/certificates/ocsp/OcspResponseGeneratorSessionBean.java
@@ -1531,9 +1531,9 @@ public class OcspResponseGeneratorSessionBean implements OcspResponseGeneratorSe
         } else {
             dataMap.put(OcspKeyBinding.PROPERTY_RESPONDER_ID_TYPE, ResponderIdType.KEYHASH.name());
         }
-        dataMap.put(OcspKeyBinding.PROPERTY_MAX_AGE, Long.valueOf(OcspConfiguration.getMaxAge(CertificateProfileConstants.CERTPROFILE_NO_PROFILE)/1000L).intValue());
+        dataMap.put(OcspKeyBinding.PROPERTY_MAX_AGE, (int)(OcspConfiguration.getMaxAge(CertificateProfileConstants.CERTPROFILE_NO_PROFILE)/1000L));
         dataMap.put(OcspKeyBinding.PROPERTY_NON_EXISTING_GOOD, Boolean.valueOf(OcspConfiguration.getNonExistingIsGood()));
-        dataMap.put(OcspKeyBinding.PROPERTY_UNTIL_NEXT_UPDATE, Long.valueOf(OcspConfiguration.getUntilNextUpdate(CertificateProfileConstants.CERTPROFILE_NO_PROFILE)/1000L).intValue());
+        dataMap.put(OcspKeyBinding.PROPERTY_UNTIL_NEXT_UPDATE, (int)(OcspConfiguration.getUntilNextUpdate(CertificateProfileConstants.CERTPROFILE_NO_PROFILE)/1000L));
         dataMap.put(OcspKeyBinding.PROPERTY_REQUIRE_TRUSTED_SIGNATURE, Boolean.valueOf(OcspConfiguration.getEnforceRequestSigning()));
         return dataMap;
     }
